requirements.txt
discord.py==2.6.4
tweepy>=4.15.0
python-dotenv==1.0.1
flask==3.0.0
gunicorn==21.2.0
